Bootstrap Carousel - Metabox Integration
The Boostrap carousel is now integrated with Metabox: Users may setup slideshows from the WordPress admin site by editing a post or page.  Accordingly the slideshow to be displayed will change on a per post/page basis.

The "items" argument to the Bootstrap carousel template is now an array of objects (rather than an array of strings) which have the following properties:

-	url: The URL of the image
-	alt: The alternate text for the image
-	title: The title of the image
-	caption: A caption for the image